added database info for adjacent regions and scores
Analyzed tweets and regions and added adjacent regions.

+++ Tweets/models.py
@@ -4,8 +4,8 @@ class Region(models.Model):
     location = models.CharField(max_length=50)
     subunit = models.IntegerField()
     pulled = models.BooleanField(default=False)
-    tweetScore = models.IntegerField(null=True)
-    finalScore = models.IntegerField(null=True)
+    tweetScore = models.FloatField(null=True)
+    finalScore = models.FloatField(null=True)
     adjacent = models.ManyToManyField("self", blank=True, null=True)
 
     def __str__(self):
@@ -16,7 +16,7 @@ class Tweet(models.Model):
     content = models.CharField(max_length=150)
     user = models.CharField(max_length=30)
     region = models.ForeignKey(Region, on_delete=models.CASCADE, default="")
-    score = models.IntegerField(null=True)
+    score = models.FloatField(null=True)
 
     def __str__(self):
         return "%s" % (self.content)
